requirements.txt
pycaption==1.0.1
imageio==2.6.1
imageio-ffmpeg==0.3.0
tqdm==4.39.0
Pillow==8.1.1
numpy==2.0.2
